🛠️ refactor: (FE) - seprate recoil atoms
