+++ frontend/src/store/feedback.atom.ts
@@ -10,3 +10,13 @@ export const feedbackIdsState = atom<number[]>({
 	key: 'feedbackIdsState',
 	default: [],
 });
+
+export const isFbClickedState = atom({
+	key: 'isFbClickedState',
+	default: false,
+});
+
+export const isFbSyncState = atom({
+	key: 'isFbSyncState',
+	default: true,
+});
